requirements.txt
llama-index==0.9.6.post2
python-dotenv==1.0.0
gradio==4.7.1
langchain==0.0.346
gigachat==0.1.8